start to organize the app files in folders related to its responsabilities
