--- BankApp/app/src/main/java/br/com/rphmelo/bankapp/StatementModel.kt
@@ -1,6 +1,0 @@
-package br.com.rphmelo.bankapp
-
-data class StatementModel( val title: String, val description: String,
-                      val date: String,  val value: Double) {
-
-}
