+++ BankApp/app/src/main/java/br/com/rphmelo/bankapp/adapter/StatementListAdapter.kt
@@ -1,14 +1,16 @@
-package br.com.rphmelo.bankapp
+package br.com.rphmelo.bankapp.adapter
 
 import android.content.Context
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import br.com.rphmelo.bankapp.R
+import br.com.rphmelo.bankapp.model.StatementModel
 import kotlinx.android.synthetic.main.item_statement_list.view.*
 
 class StatementListAdapter(private val list: List<StatementModel>,
-                           private val context: Context) : RecyclerView.Adapter<StatementListAdapter.ViewHolder>(){
+                           private val context: Context) : RecyclerView.Adapter<ViewHolder>(){
 
     override fun onCreateViewHolder(parent: ViewGroup, p1: Int): ViewHolder {
         val view = LayoutInflater.from(context).inflate(R.layout.item_statement_list, parent, false)
@@ -21,20 +23,5 @@ class StatementListAdapter(private val list: List<StatementModel>,
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
 
-    }
-
-    class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-        fun bindView(itemList: StatementModel) {
-            val title = itemView.lblItemTitle
-            val description = itemView.lblDescription
-            val date = itemView.lblDate
-            val value = itemView.lblValue
-
-            title?.text = itemList.title
-            description?.text = itemList.description
-            date?.text = itemList.date
-            value?.text = itemList.value.toString()
-        }
     }
 }
