+++ BankApp/app/src/main/res/layout/activity_currency.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CurrencyActivity">
+    tools:context=".activity.CurrencyActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appBarLayout"
